Revert "new ts"
