+++ test/utils.js
@@ -232,12 +232,9 @@ describe('Utility Methods', () => {
     });
 
     it('should return false if browser is not android', (done) => {
-        Object.defineProperty(global, 'navigator', {
-            value: {
-                userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.96 Safari/537.36',
-            },
-            writable: true,
-        });
+        global.navigator = {
+            userAgent: 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.96 Safari/537.36',
+        };
         assert.equal(utils.isAndroidBrowser(), false);
         done();
     });
